# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * COI Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Coverage {
  type: string;
  policy_number?: string;
  effective_date?: string;
  expiration_date?: string;
  each_occurrence?: string;
  aggregate?: string;
}

export interface Vendor {
  id: number;
  name: string;
  type: string;
  additional_insured: boolean;
  waiver_of_subrogation: boolean;
  certificate_holder: string;
  /** @nullable */
  email?: string | null;
  coverages: Coverage[];
  /** @nullable */
  source?: string | null;
  created_at?: string;
}

export interface VendorInput {
  name: string;
  type: string;
  additional_insured: boolean;
  waiver_of_subrogation: boolean;
  certificate_holder: string;
  /** @nullable */
  email?: string | null;
  coverages: Coverage[];
  /** @nullable */
  source?: string | null;
}

export interface CoiExtractInput {
  pdf_base64: string;
  vendor_type: string;
}

export interface CoiExtractResult {
  insured_name: string;
  certificate_holder: string;
  additional_insured: boolean;
  waiver_of_subrogation: boolean;
  coverages: Coverage[];
}

export interface Stats {
  total: number;
  compliant: number;
  expiring: number;
  non_compliant: number;
}