# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * COI Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Coverage } from './coverage';

export interface Vendor {
  id: number;
  name: string;
  type: string;
  additional_insured: boolean;
  waiver_of_subrogation: boolean;
  certificate_holder: string;
  /** @nullable */
  email?: string | null;
  coverages: Coverage[];
  /** @nullable */
  source?: string | null;
  created_at?: string;
}