# lib/db/src/schema/vendors.ts
import { pgTable, serial, text, boolean, jsonb, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const vendorsTable = pgTable("vendors", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  type: text("type").notNull(),
  additionalInsured: boolean("additional_insured").notNull().default(false),
  waiverOfSubrogation: boolean("waiver_of_subrogation").notNull().default(false),
  certificateHolder: text("certificate_holder").notNull().default(""),
  coverages: jsonb("coverages").notNull().default([]),
  source: text("source"),
  email: text("email"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertVendorSchema = createInsertSchema(vendorsTable).omit({ id: true, createdAt: true });
export type InsertVendor = z.infer<typeof insertVendorSchema>;
export type Vendor = typeof vendorsTable.$inferSelect;